// src/app/model/Workspace/OtherInformation.model.ts
export class OtherInformation {
    MaritalStatusId?: number;
    ReligionId?: number;
    Language2Id?: number
    Language3Id?: number;
    SecondaryAddressLine1: string;
    SecondaryAddressLine2: string;
    SecondaryCity: string;
    SecondaryStateCode: string;
    SecondaryZipCode: string;
    SecondaryZipPlus4: string;
    IsLivingWill?: boolean;
    AdvancedDirectivesId?: boolean;
    Comments?: string;
    Note: string;
    employerName: string;
    employerAddressLine1: string;
    employerAddressLine2: string;
    employerCity: string;
    employerState: string;
    employerZipCode: number;
    employerZipPlus4: number;
    DateCreated: string;
    CreatedByUserId: string;
    DateLastUpdated: string;
    LastUpdatedByUserId: string;
  
}