// src/app/model/Workspace/EmployerInformation.model.ts
export class EmployerInformation{
    PatientId: number;
    PatientEmployerId?: number;
    EmployerName: string;
    AddressLine1: string;
    AddressLine2: string;
    City: string;
    State: string;
    ZipCode: number;
    ZipPlus4: number;
    DateCreated: string;
    CreatedByUserId: string;
    DateLastUpdated: string;
    LastUpdatedByUserId: string;

}